会签后台管理系统和微信端php后台代码

+++ java/huiqian/yuanzhuo-huiqian-module-service/src/main/java/com/yuanzhuo/huiqian/service/impl/ActivityServiceImpl.java
@@ -7,16 +7,17 @@ import com.yuanzhuo.huiqian.dao.ActivityMapper;
 import com.yuanzhuo.huiqian.dao.FansMapper;
 import com.yuanzhuo.huiqian.dao.UserActivityFansMapper;
 import com.yuanzhuo.huiqian.model.*;
-import com.yuanzhuo.huiqian.service.ActivityService;
-import com.yuanzhuo.huiqian.service.FansMessageService;
-import com.yuanzhuo.huiqian.service.UserActivityService;
-import com.yuanzhuo.huiqian.service.WhiteListService;
+import com.yuanzhuo.huiqian.service.*;
 import com.yuanzhuo.huiqian.util.ActivityNoUtil;
 import com.yuanzhuo.huiqian.util.DateUtil;
+import jxl.Cell;
+import jxl.Sheet;
+import jxl.Workbook;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.File;
 import java.util.Date;
 import java.util.List;
 
@@ -30,10 +31,10 @@ public class ActivityServiceImpl extends BaseServiceImpl implements ActivityServ
     private ActivityMapper activityMapper;
 
     @Autowired
-    private FansMapper fansMapper;
+    private FansService fansService;
 
     @Autowired
-    private UserActivityFansMapper userActivityFansMapper;
+    private UserActivityFansService userActivityFansService;
 
     @Autowired
     private FansMessageService fansMessageService;
@@ -43,6 +44,15 @@ public class ActivityServiceImpl extends BaseServiceImpl implements ActivityServ
 
     @Autowired
     private UserActivityService userActivityService;
+
+    @Autowired
+    private ActivityFansService activityFansService;
+
+    @Autowired
+    private ActivityWinningListService activityWinningListService;
+
+    @Autowired
+    private PrizeService prizeService;
 
     @Override
     public List<Activity> getAllActivity(int pageIndex, int pageSize) throws Exception {
@@ -108,13 +118,19 @@ public class ActivityServiceImpl extends BaseServiceImpl implements ActivityServ
         if (null == activity) {
             throw new BusinessException(HuiqianErrorCode.HUIQIAN_ACTIVITYCENTER_ACTIVITY_NOT_EXISTS);
         }
+        Date beginTime = activity.getBeginTime();
+        Date currentTime = new Date();
+        if (beginTime.before(currentTime)) {
+            logger.info("当前时间:{}晚于活动开始时间:{},直接退出.", DateUtil.dateToString(currentTime), DateUtil.dateToString(beginTime));
+            return;
+        }
         int activityId = activity.getId();
         UserActivity userActivity = userActivityService.getByActivityId(activityId);
         fansMessageService.deleteMessageByActivityAndFans(activityId, 0);
         List<WhiteList> whiteLists = whiteListService.getWhiteListByActivityId(activityId);
 //        whiteListService.deleteByActivityAndFans (activityId, 0);
-        fansMapper.deleteByActivityNo(activityNo);
-        userActivityFansMapper.deleteByActivityNo(activityNo);
+        fansService.deleteByActivityNo(activityNo);
+        userActivityFansService.deleteByActivityNo(activityNo);
         int orderNum = 1;
         for (WhiteList whiteList : whiteLists) {
             Fans fans = new Fans();
@@ -125,7 +141,7 @@ public class ActivityServiceImpl extends BaseServiceImpl implements ActivityServ
             fans.setCompany(whiteList.getFans().getCompany());
             fans.setPrize(String.valueOf(whiteList.getPrizeCode()));
             fans.setCreateTime(new Date());
-            fansMapper.insertSelective(fans);
+            fansService.saveFans(fans);
 
             UserActivityFans userActivityFans = new UserActivityFans();
             userActivityFans.setFansId(fans.getId());
@@ -133,11 +149,60 @@ public class ActivityServiceImpl extends BaseServiceImpl implements ActivityServ
             userActivityFans.setUserId(userActivity.getUserId());
             userActivityFans.setOrderNum(orderNum);
             userActivityFans.setStatus(HuiqianFansStatus.HUIQIAN_USERCENTER_FANS_STATUS_PROCESS.ordinal());
-            userActivityFansMapper.insertSelective(userActivityFans);
+            userActivityFansService.saveUserActivityFans(userActivityFans);
 
             whiteList.setFansId(fans.getId());
             whiteListService.updateWhiteList(whiteList);
             orderNum++;
+        }
+    }
+
+    @Override
+    public void deleteActivityById(int activityId) {
+        activityFansService.deleteByActivityId(String.valueOf(activityId));
+        activityWinningListService.deleteByActivityId(activityId);
+        fansMessageService.deleteByActivityId(activityId);
+        prizeService.deleteByActivityId(activityId);
+        whiteListService.deleteByActivityId(activityId);
+        userActivityFansService.deleteByActivityId(activityId);
+        userActivityService.deleteByActivityId(activityId);
+        activityMapper.deleteByPrimaryKey(activityId);
+    }
+
+    @Override
+    public void recoverActivityDataByActivityNoAndFile(int userId, int activityId, String filePath) {
+        Workbook workbook = null;
+        String currentName = "";
+        try {
+            workbook = Workbook.getWorkbook(new File(filePath));
+            Sheet sheet = workbook.getSheet(0);
+            int totalCount = sheet.getRows();
+            System.out.println("总共有多少条数据:" + totalCount);
+            for (int i = 0; i < totalCount; i++) {
+                Cell[] cells = sheet.getRow(i);
+                String nick = cells[0].getContents();
+                String name = cells[1].getContents();
+                String mobile = cells[2].getContents();
+                String company = cells[3].getContents();
+                currentName = nick;
+                Fans fans = new Fans();
+                fans.setNick(nick);
+                fans.setName(name);
+                fans.setPhone(mobile);
+                fans.setCompany(company);
+                fans.setCreateTime(new Date());
+                fansService.saveFans(fans);
+                UserActivityFans userActivityFans = new UserActivityFans();
+                userActivityFans.setActivityId(activityId);
+                userActivityFans.setFansId(fans.getId());
+                userActivityFans.setStatus(HuiqianFansStatus.HUIQIAN_USERCENTER_FANS_STATUS_PROCESS.ordinal());
+                userActivityFans.setOrderNum(i + 1);
+                userActivityFans.setUserId(userId);
+                userActivityFansService.saveUserActivityFans(userActivityFans);
+                logger.info("插入用户名为:{}成功!", nick);
+            }
+        } catch (Exception e) {
+            logger.error("当前用户名为:{},异常信息:{}", currentName, e.getMessage());
         }
     }
